fix code organization with namespace.  Remove railtie as it's no longer needed
